Changed models from embedded to reference format
